% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignment}
\alias{assignment}
\alias{[<-,dgRMatrix,index,index,replValue-method}
\alias{[<-,dgRMatrix,missing,index,replValue-method}
\alias{[<-,dgRMatrix,index,missing,replValue-method}
\alias{[<-,dgRMatrix,missing,missing,replValue-method}
\alias{[<-,dgRMatrix,index,index,sparseVector-method}
\alias{[<-,dgRMatrix,missing,index,sparseVector-method}
\alias{[<-,dgRMatrix,index,missing,sparseVector-method}
\alias{[<-,dgRMatrix,missing,missing,sparseVector-method}
\alias{[<-,dgRMatrix,nsparseVector,index,replValue-method}
\alias{[<-,dgRMatrix,nsparseVector,missing,replValue-method}
\alias{[<-,dgRMatrix,index,nsparseVector,replValue-method}
\alias{[<-,dgRMatrix,missing,nsparseVector,replValue-method}
\alias{[<-,dgRMatrix,nsparseVector,nsparseVector,replValue-method}
\alias{[<-,dgRMatrix,nsparseVector,index,sparseVector-method}
\alias{[<-,dgRMatrix,nsparseVector,missing,sparseVector-method}
\alias{[<-,dgRMatrix,index,nsparseVector,sparseVector-method}
\alias{[<-,dgRMatrix,missing,nsparseVector,sparseVector-method}
\alias{[<-,dgRMatrix,nsparseVector,nsparseVector,sparseVector-method}
\alias{[<-,dgRMatrix,lsparseVector,index,replValue-method}
\alias{[<-,dgRMatrix,lsparseVector,missing,replValue-method}
\alias{[<-,dgRMatrix,index,lsparseVector,replValue-method}
\alias{[<-,dgRMatrix,missing,lsparseVector,replValue-method}
\alias{[<-,dgRMatrix,lsparseVector,lsparseVector,replValue-method}
\alias{[<-,dgRMatrix,lsparseVector,index,sparseVector-method}
\alias{[<-,dgRMatrix,lsparseVector,missing,sparseVector-method}
\alias{[<-,dgRMatrix,index,lsparseVector,sparseVector-method}
\alias{[<-,dgRMatrix,missing,lsparseVector,sparseVector-method}
\alias{[<-,dgRMatrix,lsparseVector,lsparseVector,sparseVector-method}
\alias{[<-,dgRMatrix,lsparseMatrix,index,sparseMatrix-method}
\alias{[<-,dgRMatrix,lsparseMatrix,missing,sparseMatrix-method}
\alias{[<-,dgRMatrix,index,lsparseMatrix,sparseMatrix-method}
\alias{[<-,dgRMatrix,missing,lsparseMatrix,sparseMatrix-method}
\alias{[<-,dgRMatrix,lsparseMatrix,lsparseMatrix,sparseMatrix-method}
\title{Assignment operator for CSR matrices}
\usage{
\S4method{[}{dgRMatrix,index,index,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,missing,index,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,index,missing,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,missing,missing,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,index,index,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,missing,index,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,index,missing,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,missing,missing,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,nsparseVector,index,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,nsparseVector,missing,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,index,nsparseVector,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,missing,nsparseVector,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,nsparseVector,nsparseVector,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,nsparseVector,index,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,nsparseVector,missing,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,index,nsparseVector,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,missing,nsparseVector,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,nsparseVector,nsparseVector,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,lsparseVector,index,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,lsparseVector,missing,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,index,lsparseVector,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,missing,lsparseVector,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,lsparseVector,lsparseVector,replValue}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,lsparseVector,index,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,lsparseVector,missing,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,index,lsparseVector,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,missing,lsparseVector,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,lsparseVector,lsparseVector,sparseVector}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,lsparseMatrix,index,sparseMatrix}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,lsparseMatrix,missing,sparseMatrix}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,index,lsparseMatrix,sparseMatrix}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,missing,lsparseMatrix,sparseMatrix}(x, i, j, ...) <- value

\S4method{[}{dgRMatrix,lsparseMatrix,lsparseMatrix,sparseMatrix}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A CSR matrix whose values are to be replaced.}

\item{i}{The indices of the rows to replace.}

\item{j}{The indices of the columns to replace.}

\item{...}{Not used}

\item{value}{The values to replace with.}
}
\value{
The same `x` input with the values `[i,j]` set to `value`.
If the result is a full matrix (e.g. `x[,] <- 1`), the object will
be a dense matrix from base R.
}
\description{
Assign values to a CSR matrix.
Note: this will only be a relatively fast operation when
assigning contiguous row sequences. Only some of the potential
assignment cases to a CSR matrix are replaced here - for example,
cases that involve uneven recycling of vectors will be left to
the `Matrix` package.
}
\examples{
library(Matrix)
library(MatrixExtra)
set.seed(1)
X <- rsparsematrix(5, 3, .5, repr="R")
X[1:3] <- 0
X
}
