% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deepcopy_sparse_object}
\alias{deepcopy_sparse_object}
\title{Deep copy sparse matrices and vectors}
\usage{
deepcopy_sparse_object(X)
}
\arguments{
\item{X}{A sparse matrix or sparse vector from the `Matrix` package.}
}
\value{
The same input `X` with the fields replaced with deep copies.
}
\description{
Generates a deep copy of a sparse matrix or sparse vector
object, which can come useful when the matrix is to later be passed to
functions that will potentially modify it in-place, such as \link{sort_sparse_indices}.
}
