% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.R
\name{cbind2-method}
\alias{cbind2-method}
\alias{cbind2,TsparseMatrix,TsparseMatrix-method}
\alias{cbind2,TsparseMatrix,sparseVector-method}
\alias{cbind2,sparseVector,TsparseMatrix-method}
\alias{cbind2,CsparseMatrix,sparseVector-method}
\alias{cbind2,sparseVector,CsparseMatrix-method}
\alias{cbind2,sparseVector,sparseVector-method}
\title{Concatenate COO matrices by columns}
\usage{
\S4method{cbind2}{TsparseMatrix,TsparseMatrix}(x, y)

\S4method{cbind2}{TsparseMatrix,sparseVector}(x, y)

\S4method{cbind2}{sparseVector,TsparseMatrix}(x, y)

\S4method{cbind2}{CsparseMatrix,sparseVector}(x, y)

\S4method{cbind2}{sparseVector,CsparseMatrix}(x, y)

\S4method{cbind2}{sparseVector,sparseVector}(x, y)
}
\arguments{
\item{x}{First matrix to concatenate.}

\item{y}{Second matrix to concatenate.}
}
\value{
A COO matrix.
}
\description{
`cbind2` method for the sparse COO matrix and sparse vector classes from `Matrix`,
taking the most efficient route for the concatenation according to the input types.
}
\examples{
library(Matrix)
library(MatrixExtra)
set.seed(1)
X <- rsparsematrix(3, 4, .3)
X <- as(X, "TsparseMatrix")
inherits(cbind2(X, X), "dgTMatrix")
}
