% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.R
\name{rbind2-method}
\alias{rbind2-method}
\alias{rbind2,RsparseMatrix,RsparseMatrix-method}
\alias{rbind2,sparseVector,RsparseMatrix-method}
\alias{rbind2,RsparseMatrix,sparseVector-method}
\alias{rbind2,sparseVector,sparseVector-method}
\alias{rbind2,CsparseMatrix,CsparseMatrix-method}
\alias{rbind2,sparseVector,CsparseMatrix-method}
\alias{rbind2,CsparseMatrix,sparseVector-method}
\alias{rbind2,RsparseMatrix,CsparseMatrix-method}
\alias{rbind2,CsparseMatrix,RsparseMatrix-method}
\alias{rbind2,RsparseMatrix,TsparseMatrix-method}
\alias{rbind2,TsparseMatrix,RsparseMatrix-method}
\title{Concatenate CSR matrices by rows}
\usage{
\S4method{rbind2}{RsparseMatrix,RsparseMatrix}(x, y)

\S4method{rbind2}{sparseVector,RsparseMatrix}(x, y)

\S4method{rbind2}{RsparseMatrix,sparseVector}(x, y)

\S4method{rbind2}{sparseVector,sparseVector}(x, y)

\S4method{rbind2}{CsparseMatrix,CsparseMatrix}(x, y)

\S4method{rbind2}{sparseVector,CsparseMatrix}(x, y)

\S4method{rbind2}{CsparseMatrix,sparseVector}(x, y)

\S4method{rbind2}{CsparseMatrix,CsparseMatrix}(x, y)

\S4method{rbind2}{RsparseMatrix,CsparseMatrix}(x, y)

\S4method{rbind2}{CsparseMatrix,RsparseMatrix}(x, y)

\S4method{rbind2}{RsparseMatrix,TsparseMatrix}(x, y)

\S4method{rbind2}{TsparseMatrix,RsparseMatrix}(x, y)
}
\arguments{
\item{x}{First matrix to concatenate.}

\item{y}{Second matrix to concatenate.}
}
\value{
A CSR matrix (`dgRMatrix` or `ndRMatrix` depending on the inputs).
}
\description{
`rbind2` method for the `dgRMatrix`, `ngRMatrix`, and `sparseVector` classes
from the `Matrix` package. This method will concatenate CSR (a.k.a. RsparseMatrix)
matrix objects without converting them to triplets/COO in the process, and will also
concatenate sparse vectors assuming they are row vectors.
}
\examples{
library(Matrix)
library(MatrixExtra)
set.seed(1)
X <- rsparsematrix(3, 4, .3)
X <- as(X, "RsparseMatrix")
inherits(rbind2(X, X), "dgRMatrix")
inherits(rbind(X, X, as.csc.matrix(X), X), "dgRMatrix")
}
